ai-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
groq==0.4.2
chromadb==0.4.22
sentence-transformers==2.2.2
python-dotenv==1.0.0
gunicorn==21.2.0
